Fix async test configuration and optimize integration tests
- Fix pytest-asyncio configuration for CI compatibility
- Optimize integration test performance (10+ min -> ~2 min)
- Fix target variable type issues in classification tests
- Remove unused imports and clean up code
- Add pytest.ini for better async test support
- Update test requirements for better compatibility

+++ src/aml_agent/utils.py
@@ -507,8 +507,8 @@ def create_sample_data(
 
     # Generate target
     if task_type == TaskType.CLASSIFICATION:
-        # Binary classification
-        target = np.random.randint(0, 2, n_samples).astype(np.float64)
+        # Binary classification - use int64 for proper classification
+        target = np.random.randint(0, 2, n_samples).astype(np.int64)
     else:
         # Regression
         target = np.random.randn(n_samples).astype(np.float64)
